+++ src/App.tsx
@@ -12,11 +12,11 @@ import PretendardMedium from '@/assets/fonts/Pretendard-Medium.woff';
 import Home from './pages/Home';
 import MyEstimatePage from './pages/Home/MyEstimate';
 import AllEstimates from './pages/Home/AllEstimates';
-import My from './pages/Expert/My';
 import Estimates from './pages/Expert/Estimates';
 import Estimate from './pages/Expert/Estimate';
 import ExpertMyEstimate from './pages/Expert/MyEstimate';
-import MyEdit from './pages/Expert/MyEdit';
+import PortfolioEdit from './pages/Expert/PortfolioEdit';
+import PortfolioEditor from './pages/Expert/PortfolioEditor';
 
 import Agreement from './pages/Auth/Agreement';
 import Signup from './pages/Auth/Signup';
@@ -31,13 +31,14 @@ import ChatRoom from './pages/chat/ChatRoom';
 import MenuBar from './pages/MenuBar';
 import CommunityPage from './pages/Home/Community';
 import CheckMyEstimatePage from './pages/Home/EstimateCheckPage';
+import ExpertProfile from './pages/Expert/Profile';
+import Portfolio from './pages/Expert/Portfolio';
 
 // 컴포넌트
 import LoginModal from './components/LoginModal';
 import useAuthStore from './store/useAuthStore';
 import { Navbar } from './components/Navbar';
 import { Footer } from './components/Footer';
-import ExpertProfile from './pages/Expert/Profile';
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -73,11 +74,12 @@ const AppRoutes = () => {
         <Route path="/" element={<Home />} />
         <Route path="/MyEstimate" element={<MyEstimatePage />} />
         <Route path="/expert/profile" element={<ExpertProfile />} />
-        <Route path="/expert/my" element={<My />} />
-        <Route path="/expert/my/edit" element={<MyEdit />} />
+        <Route path="/expert/portfolio/edit" element={<PortfolioEdit />} />
+        <Route path="/expert/portfolio/editor" element={<PortfolioEditor />} />
+        <Route path="/expert/:userID" element={<Portfolio />} />
         <Route path="/expert/estimates" element={<Estimates />} />
         <Route path="/expert/estimate" element={<Estimate />} />
-        <Route path="/expert/myestimate" element={<ExpertMyEstimate />} />
+        <Route path="/expert/my/estimate" element={<ExpertMyEstimate />} />
         <Route path="/expert-register" element={<Register />} />
         <Route path="/detail-category/:categoryId" element={<DetailCategory />} />
         <Route path="/location" element={<Location />} />
