+++ src/components/modals/Expert/Portfolio.modal.tsx
@@ -1,7 +1,7 @@
 import { Modal } from '@/components/Modal';
 import { useEditMyProfileModalStore } from '@/store/modals/useExpertModalStore';
 
-import * as M from '@/styles/components/modals/Expert/My.style';
+import * as M from '@/styles/components/modals/Expert/Portfolio.style';
 
 const Check = () => (
   <svg xmlns="http://www.w3.org/2000/svg" width="14" height="15" viewBox="0 0 14 15" fill="none">
