+++ src/components/modals/Expert/PortfolioEdit.modal.tsx
@@ -1,7 +1,7 @@
 import { Modal } from '@/components/Modal';
 import { useEditMyPortfolioModalStore } from '@/store/modals/useExpertModalStore';
 
-import * as ME from '@/styles/components/modals/Expert/MyEdit.style';
+import * as ME from '@/styles/components/modals/Expert/PortfolioEdit.style';
 
 const Check = () => (
   <svg xmlns="http://www.w3.org/2000/svg" width="14" height="15" viewBox="0 0 14 15" fill="none">
