+++ src/pages/Expert/PortfolioEditor.tsx
@@ -1,0 +1,3 @@
+export default function PortfolioEditor() {
+  return <>d</>;
+}
